feat(payment): PI-2777 Move itemsRequireShipping method to the utils
