+++ packages/payment-integration-api/src/cart/index.ts
@@ -11,4 +11,3 @@ export {
     LineItem,
 } from './line-item';
 export { default as LineItemMap } from './line-item-map';
-export { getPhysicalItem } from './line-items.mock';
