+++ packages/payment-integration-api/src/mocks/line-items.mock.ts
@@ -1,4 +1,4 @@
-import { CustomItem, DigitalItem, GiftCertificateItem, PhysicalItem } from './line-item';
+import { CustomItem, DigitalItem, GiftCertificateItem, PhysicalItem } from '../cart';
 
 export function getCustomItem(): CustomItem {
     return {
